Add files via upload
YAMNet (Yet Another Mobile Network) is a pre-trained deep neural network that can predict audio events from 521 classes, such as laughter, barking, or a siren. This Project has been trained to identify and plot the sound of a cat. It plots the necessary charts to illustrate the frequency and the most corresponding animal to which the sound belongs i.e. a Cat.